style deadline component
